+++ src/operator/numpy/random/np_multinomial_op.h
